Correcting invalid usage of reflect.Type
When `checkType.Elem()` is called on any type that is not `Array, Chan, Map, Ptr, or Slice`, it will panic. The else statement that has been corrected, is specifically not dealing with a pointer. As we're dealing with controllers, I expect this should be dealing with an instance of a struct? (too new to Go to know)

Fixes #1336